Fix eager loading for HABTM association
There is a bug in Rails 3.2.x and BINARY column that Rails does not know
it should convert BINARY string into an UUID object first before trying
to cache into association cache. This patch make it so that those ids
that are used for caching are converted to UUID object as they are
supposed to.